// src/Models/ClientsenAnomalieModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */

export interface ClientsenAnomalie {
  ID?: number;
  Title?: string;
  Matricule?: string;
  EmailGFC?: string;
  Nomagence?: string;
  Montant?: number;
  P_x00e9_riodicit_x00e9_?: string;
  Date?: string;
  "StatutAction#Id"?: number;
  StatutAction?: Record<string, unknown>;
  TypeClient?: string;
  Typedossier?: string;
  TotalSains?: number;
  TotalTr_x00e9_sorerie?: number;
  TotalNPL?: number;
  NomGFC?: string;
  Modified?: string;
  Created?: string;
  Numerodossier?: string;
  "Author#Claims"?: string;
  Author?: Record<string, unknown>;
  "Editor#Claims"?: string;
  Editor?: Record<string, unknown>;
  R_x00e9_seau?: string;
  OData__ColorTag?: string;
  ComplianceAssetId?: string;
  "{Identifier}"?: string;
  "{IsFolder}"?: boolean;
  "{Thumbnail}"?: Record<string, unknown>;
  "{Link}"?: string;
  "{Name}"?: string;
  "{FilenameWithExtension}"?: string;
  "{Path}"?: string;
  "{FullPath}"?: string;
  "{ModerationStatus}"?: string;
  "{ModerationComment}"?: string;
  "{ContentType}#Id"?: string;
  "{ContentType}"?: Record<string, unknown>;
  "{HasAttachments}"?: boolean;
  "{Attachments}@odata.type"?: string;
  "{Attachments}"?: unknown;
  "{VersionNumber}"?: string;
  "{TriggerWindowStartToken}"?: string;
  "{TriggerWindowEndToken}"?: string;
}